duckiebot_cs_zoo/requirements.txt
picamera
PyTurboJPEG
opencv-python==4.4.0.46
joblib
numpy
pyserial==3.4
